pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.4",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "jdatetime>=5.2.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot>=22.3",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "ta-lib>=0.6.5",
    "trafilatura>=2.0.0",
    "werkzeug>=3.1.3",
]
